+++ src/main/java/roomescape/exception/ReservationTimeException.java
@@ -1,9 +1,10 @@
-package roomescape.core.exception;
+package roomescape.exception;
 
 import org.springframework.http.HttpStatus;
+import roomescape.core.exception.CustomException;
 
 public class ReservationTimeException extends CustomException {
-    
+
     public ReservationTimeException(HttpStatus httpStatus, String message) {
         super(httpStatus, message);
     }
